feat: complete AllProducts component
