--- client/components/Footer.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import {Link} from 'react-router';
-
-export default Footer = () => {
-  return (
-    <div>
-      <img />
-      <Link to="">Home</Link>
-    </div>
-  )
-}
-
