+++ client/index.js
@@ -1,22 +1,22 @@
-import {Route, Switch} from 'react-router';
-import React from 'react';
-
+import React, {Component} from 'react'
+import {Route, Switch} from 'react-router-dom';
+import AllProducts from './components/AllProducts';
+import {fetchAllProducts} from './reducers';
+import store from './store';
 // import navbar, footer, all Components
 
-export default Index = function(){
-  return (
-    <div>
-      <Navbar />
-        <Switch>
-          <Route exact path="/" component={AllProducts}/>
-          <Route path="/products/:productId" component={SingleProduct}/>
-          <Route exact path="/profile" component={UserProfile}/>
-          <Route path="/cart" component={Cart} />
-          <Route path="/checkout" component={Checkout} />
-          <Route path="/order/:orderId" component={SingleOrder} />
-          <Route component={NotFound} />
-        </Switch>
-      <Footer />
-    </div>
-  )
+export default class Index extends Component {
+  componentDidMount() {
+    store.dispatch(fetchAllProducts())
+  }
+
+  render() {
+    return (
+      <div>
+          <Switch>
+            <Route exact path="/" component={AllProducts} />
+          </Switch>
+      </div>
+    )
+  }
 }
