+++ package.json
@@ -20,6 +20,7 @@
   },
   "homepage": "https://github.com/GH1707-WorkingName/GraceShopper#readme",
   "dependencies": {
+    "axios": "^0.16.2",
     "body-parser": "^1.18.0",
     "connect-session-sequelize": "^4.1.0",
     "express": "^4.15.4",
