+++ server/api/index.js
@@ -1,3 +1,5 @@
+'use strict'
+
 const router = require('express').Router()
 
 router.use('/products', require('./products'))
