+++ server/api/products.js
@@ -8,3 +8,5 @@ router.get('/', (req, res, next) => {
     })
     .catch(next)
 })
+
+module.exports = router;
